refactor(user): change from vuex to pinia
