+++ apps/web/src/store/index.ts
@@ -3,7 +3,6 @@ import Vuex from 'vuex';
 
 import display from './modules/display';
 // eslint-disable-next-line import/no-cycle
-import user from './modules/user';
 // eslint-disable-next-line import/no-cycle
 import plugins from './plugins';
 
@@ -11,7 +10,6 @@ Vue.use(Vuex);
 
 const store = new Vuex.Store<any>({
     modules: {
-        user,
         display,
     },
     plugins,
