+++ apps/web/src/store/plugins.ts
@@ -1,13 +1,11 @@
 import createLogger from 'vuex/dist/logger';
 
 import displayPlugins from '@/store/modules/display/plugins';
-// eslint-disable-next-line import/no-cycle
-import userPlugins from '@/store/modules/user/plugins';
+
 
 const DEBUG = import.meta.env.DEV;
 
 const PLUGINS = [
-    ...userPlugins,
     ...displayPlugins,
 ];
