+++ apps/web/src/store/util.ts
@@ -1,9 +1,7 @@
 import { LocalStorageAccessor } from '@cloudforet/core-lib/local-storage-accessor';
 
-import { store } from '@/store/index';
 
-export const initServiceSettingsStore = <ServiceSettingStates>(service: string): ServiceSettingStates | undefined => {
-    const userId = store.state.user.userId;
+export const initServiceSettingsStore = <ServiceSettingStates>(service: string, userId?: string): ServiceSettingStates | undefined => {
     if (userId) {
         const userSettings = LocalStorageAccessor.getItem(userId) ?? {};
         return userSettings[service] || {};
